Abstract node builders similar to char and export
