+++ src/js/decorations/hard-break.js
@@ -1,15 +1,5 @@
-import createDeco from '../utils/create-deco';
+import node from './node';
 
 export default (
   predicate = node => node.type === node.type.schema.nodes.hard_break
-) => {
-  return (from, to, doc, decos) => {
-    let newDecos = decos;
-    doc.nodesBetween(from, to, (node, pos) => {
-      newDecos = predicate(node)
-        ? newDecos.add(doc, [createDeco(pos, 'break')])
-        : newDecos;
-    });
-    return newDecos;
-  };
-};
+) => node('break', (_, pos) => pos)(predicate);
