+++ src/js/index.js
@@ -29,8 +29,10 @@ export default builders => {
   });
 };
 
-export { default as space } from './decorations/space';
 export { default as character } from './decorations/character';
+export { default as node } from './decorations/node';
+
+export { default as space } from './decorations/space';
 export { default as hardBreak } from './decorations/hard-break';
 export { default as paragraph } from './decorations/paragraph';
